Fix: add shellcheck of all .sh files in CI. Fix: add the missing shebang and add set -e in all .sh files. Fix: now submodules initializing only if they were not inited before. Fix: verilator installing only if it was not installed before. Fix: now toolchain not exports in path in dev version. Fix: deleted finding libs for verilator.

+++ CI/load_toolchain.sh
@@ -1,3 +1,7 @@
-wget https://se.math.spbu.ru/riscv-public/sc-dt-2023.08/riscv-gcc.tar.xz
-tar -xf riscv-gcc.tar.xz
-rm riscv-gcc.tar.xz
+#!/bin/sh
+set -e
+if [ ! -e "./riscv-gcc/bin" ]; then
+	wget https://se.math.spbu.ru/riscv-public/sc-dt-2023.08/riscv-gcc.tar.xz
+	tar -xf riscv-gcc.tar.xz
+	rm riscv-gcc.tar.xz
+fi
